requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
Django==5.0.3
entmax==1.3
executing==2.0.1
filelock==3.13.4
Flask==3.0.1
flatbuffers==24.3.25
fonttools==4.47.2
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
gTTS==2.5.1
h5py==3.11.0
idna==3.6
imageio==2.34.1
ipykernel==6.28.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.5.9
networkx==3.3
nltk==3.8.1
ntscraper==0.3.10
numpy==1.26.3
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
rich==13.7.1
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.1
secure-smtplib==0.1.1
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
textblob==0.17.1
tifffile==2024.4.18
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
